+++ Sources/Client/Model/Event.swift
@@ -66,7 +66,7 @@ public enum Event: Decodable, Equatable {
     case channelHidden(HiddenChannelResponse, ChannelId?, EventType)
     
     /// When a new message was added on a channel (when watching the channel).
-    case messageNew(Message, _ watcherCount: Int, ChannelId?, EventType)
+    case messageNew(Message, UnreadCount, _ watcherCount: Int, ChannelId?, EventType)
     /// When a message was updated (when watching the channel).
     case messageUpdated(Message, ChannelId?, EventType)
     /// When a message was deleted (when watching the channel).
@@ -125,7 +125,7 @@ public enum Event: Decodable, Equatable {
              .channelHidden(_, _, let type),
              
              .messageRead(_, _, let type),
-             .messageNew(_, _, _, let type),
+             .messageNew(_, _, _, _, let type),
              .messageDeleted(_, _, _, let type),
              .messageUpdated(_, _, let type),
              
@@ -175,7 +175,7 @@ public enum Event: Decodable, Equatable {
              .channelHidden(_, let cid, _),
              
              .messageRead(_, let cid, _),
-             .messageNew(_, _, let cid, _),
+             .messageNew(_, _, _, let cid, _),
              .messageDeleted(_, _, let cid, _),
              .messageUpdated(_, let cid, _),
              
@@ -232,7 +232,7 @@ public enum Event: Decodable, Equatable {
         case .memberAdded(let member, _, _),
              .memberUpdated(let member, _, _):
             return member.user
-        case .messageNew(let message, _, _, _),
+        case .messageNew(let message, _, _, _, _),
              .notificationMessageNew(let message, _, _, _, _):
             return message.user
         default:
@@ -331,7 +331,7 @@ public enum Event: Decodable, Equatable {
         // Message
         case .messageNew:
             let watcherCount = try container.decodeIfPresent(Int.self, forKey: .watcherCount) ?? 0
-            self = try .messageNew(message(), watcherCount, cid(), type)
+            self = try .messageNew(message(), unreadCount(), watcherCount, cid(), type)
         case .messageRead:
             let unreadMessages = try container.decodeIfPresent(Int.self, forKey: .unreadMessagesCount) ?? 0
             let messageRead = try MessageRead(user: user(), lastReadDate: created(), unreadMessagesCount: unreadMessages)